연차VO int->double타입으로 변환
화면에 보이는 소수점은 처리할예정

+++ sevenFS/src/main/java/kr/or/ddit/sevenfs/controller/organization/DclzTypeController.java
@@ -208,7 +208,7 @@ public class DclzTypeController {
 		VacationVO emplVacation = dclztypeService.emplVacationCnt(emplNo);
 		
 		// 잔여연차 계산된값 확인
-		int remain = emplVacation.getYrycRemndrDaycnt();
+		double remain = emplVacation.getYrycRemndrDaycnt();
 		//log.info("controller잔여연차 : " + remain);
 		vacationVO.setYrycRemndrDaycnt(remain);
